requirements.txt
requests
python-dotenv
jmespath
